Restructured line search conditions module
`conditions` is now a module and each condition is in a separate file.
Documentation was improved and doc tests added. The method names and
method parameters of the `LineSearchCondition` trait were changed to be
more intuitive.
